Remove fs import from base.js
