+++ src/base.js
@@ -1,5 +1,4 @@
 import r from '@pdf-lib/restructure';
-import fs from 'fs';
 
 const formats = [];
 
@@ -24,32 +23,6 @@ const fontkit = {
     }
     throw new Error('Unknown font format');
   },
-};
-
-fontkit.openSync = function(filename, postscriptName) {
-  let buffer = fs.readFileSync(filename);
-  return fontkit.create(buffer, postscriptName);
-};
-
-fontkit.open = function(filename, postscriptName, callback) {
-  if (typeof postscriptName === 'function') {
-    callback = postscriptName;
-    postscriptName = null;
-  }
-
-  fs.readFile(filename, function(err, buffer) {
-    if (err) { return callback(err); }
-
-    try {
-      var font = fontkit.create(buffer, postscriptName);
-    } catch (e) {
-      return callback(e);
-    }
-
-    return callback(null, font);
-  });
-
-  return;
 };
 
 export default fontkit;
